# SmartHelpDeskSwagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Smart Helpdesk CRM — Swagger đầy đủ",
    "version": "1.0.0",
    "description": "Đối chiếu source ngày 2026-09-12. 25 handler nghiệp vụ / 29 method-path (4 alias auth). JSON UTF-8; UUID dạng chuỗi; Instant ISO-8601 UTC. ResponseGeneral bỏ field null ở cấp envelope. HTTP 201 vẫn có body.status=200. Lỗi Spring Security filter không bảo đảm envelope ApiError. AI chạy RuleBasedLlmAdapter nội bộ, không gọi API LLM bên ngoài."
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local; đổi host/port theo môi trường"
    }
  ],
...
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/members": {
      "post": {
        "tags": [
          "Members"
        ],
        "summary": "Đăng ký member",
        "operationId": "registerMember",
        "description": "Quyền: Public. Cùng nghiệp vụ với /api/auth/register.",
        "x-access": "Public",
        "responses": {
          "201": {
            "description": "Tạo thành công; body.status=200.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/registerMemberResponse"
                },
                "example": {
                  "status": 200,
                  "message": "Member registered successfully",
                  "data": {
                    "id": "0198f101-1234-7000-8000-000000000001",
                    "email": "client@example.com",
                    "name": "Nguyễn Văn An",
                    "role": "ADMIN",
                    "phone": "Ví dụ",
                    "companyId": "COMPANY001",
                    "companyName": "Công ty mẫu",
                    "status": "active",
                    "isDeleted": false,
                    "createdAt": "2026-09-12T03:00:00Z"
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "404": {
            "$ref": "#/components/responses/Error404"
          },
          "500": {
            "$ref": "#/components/responses/Error500"
          }
        },
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              },
              "example": {
                "email": "client@example.com",
                "password": "Example123!",
                "name": "Nguyễn Văn An",
                "role": "CLIENT",
                "phone": "0900000000",
                "companyId": "COMPANY001"
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Members"
        ],
        "summary": "Danh sách member",
        "operationId": "listMembers",
        "description": "Quyền: ADMIN. Không phân trang; findAll gồm cả member có isDeleted=true.",
        "x-access": "ADMIN",
        "responses": {
          "200": {
            "description": "Thành công.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/listMembersResponse"
                },
                "example": {
                  "status": 200,
                  "message": "success",
                  "data": []
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/SecurityDenied"
          },
          "500": {
            "$ref": "#/components/responses/Error500"
          }
        }
      }
    },
    "/api/members/{id}": {
      "get": {
        "tags": [
          "Members"
        ],
        "summary": "Chi tiết member",
        "operationId": "getMember",
        "description": "Quyền: Đã đăng nhập. Không kiểm tra chủ sở hữu hoặc cùng công ty.",
        "x-access": "Đã đăng nhập",
        "responses": {
          "200": {
            "description": "Thành công.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/getMemberResponse"
                },
                "example": {
                  "status": 200,
                  "message": "success",
                  "data": {
                    "id": "0198f101-1234-7000-8000-000000000001",
                    "email": "client@example.com",
                    "name": "Nguyễn Văn An",
                    "role": "ADMIN",
                    "phone": "Ví dụ",
                    "companyId": "COMPANY001",
                    "companyName": "Công ty mẫu",
                    "status": "active",
                    "isDeleted": false,
                    "createdAt": "2026-09-12T03:00:00Z"
                  }
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/SecurityDenied"
          },
          "404": {
            "$ref": "#/components/responses/Error404"
          },
          "500": {
            "$ref": "#/components/responses/Error500"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID tài nguyên UUID.",
            "schema": {
              "type": "string",
              "format": "uuid",
              "example": "0198f101-1234-7000-8000-000000000001"
            }
          }
        ]
      }
    },
    "/api/companies": {
      "get": {
        "tags": [
          "Companies"
        ],
        "summary": "Danh sách công ty",
        "operationId": "listCompanies",
        "description": "Quyền: Public. Sắp xếp companyName rồi ID tăng dần; không phân trang.",
        "x-access": "Public",
        "responses": {
          "200": {
            "description": "Thành công.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/listCompaniesResponse"
                },
                "example": {
                  "status": 200,
                  "message": "success",
                  "data": []
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/Error500"
          }
        },
        "security": []
      }
    },
    "/api/companies/{companyId}": {
      "get": {
        "tags": [
          "Companies"
        ],
        "summary": "Chi tiết công ty",
        "operationId": "getCompany",
        "description": "Quyền: Đã đăng nhập.",
        "x-access": "Đã đăng nhập",
        "responses": {
          "200": {
            "description": "Thành công.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/getCompanyResponse"
                },
                "example": {
                  "status": 200,
                  "message": "success",
                  "data": {
                    "companyId": "COMPANY001",
                    "companyName": "Công ty mẫu",
                    "address": "Ví dụ",
                    "phone": "Ví dụ"
                  }
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/SecurityDenied"
          },
          "404": {
            "$ref": "#/components/responses/Error404"
          },
          "500": {
            "$ref": "#/components/responses/Error500"
          }
        },
        "parameters": [
          {
            "name": "companyId",
            "in": "path",
            "required": true,
            "description": "Mã công ty, không phải UUID.",
            "schema": {
              "type": "string",
              "example": "COMPANY001"
            }
          }
        ]
      }
    },
...
    "/api/chat/inbox": {
      "get": {
        "tags": [
          "Chat"
        ],
        "summary": "Hộp thư chat",
        "operationId": "chatInbox",
        "description": "Quyền: Đã đăng nhập. Mọi tin thuộc request có quyền xem (admin/chủ request/developer được giao), gồm DB và bộ nhớ; mới nhất trước; không phân trang, không gộp hội thoại.",
        "x-access": "Đã đăng nhập",
        "responses": {
          "200": {
            "description": "Thành công.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/chatInboxResponse"
                },
                "example": {
                  "status": 200,
                  "message": "success",
                  "data": []
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/SecurityDenied"
          },
          "500": {
            "$ref": "#/components/responses/Error500"
          }
        }
      }
    },
    "/api/chat/requests/{requestId}/messages": {
      "get": {
        "tags": [
          "Chat"
        ],
        "summary": "Lịch sử chat",
        "operationId": "chatHistory",
        "description": "Quyền: ADMIN hoặc client tạo request hoặc developer được giao. Gồm DB và buffer, createdAt tăng dần.",
        "x-access": "ADMIN hoặc client tạo request hoặc developer được giao",
        "responses": {
          "200": {
            "description": "Thành công.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/chatHistoryResponse"
                },
                "example": {
                  "status": 200,
                  "message": "success",
                  "data": []
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "403": {
            "$ref": "#/components/responses/SecurityDenied"
          },
          "404": {
            "$ref": "#/components/responses/Error404"
          },
          "500": {
            "$ref": "#/components/responses/Error500"
          }
        },
        "parameters": [
          {
            "name": "requestId",
            "in": "path",
            "required": true,
            "description": "ID tài nguyên UUID.",
            "schema": {
              "type": "string",
              "format": "uuid",
              "example": "0198f101-1234-7000-8000-000000000001"
            }
          }
        ]
      },
      "post": {
        "tags": [
          "Chat"
        ],
        "summary": "Gửi tin nhắn",
        "operationId": "sendMessage",
        "description": "Quyền: ADMIN hoặc client tạo request hoặc developer được giao. HTTP 200, persisted=false. Trim message. Buffer theo request; sau 10 phút không gửi mới, scheduler kiểm tra mỗi 30 giây để lưu DB. Tin chưa lưu mất khi restart.",
        "x-access": "ADMIN hoặc client tạo request hoặc developer được giao",
        "responses": {
          "200": {
            "description": "Thành công.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/sendMessageResponse"
                },
                "example": {
                  "status": 200,
                  "message": "success",
                  "data": {
                    "id": "0198f101-1234-7000-8000-000000000001",
                    "requestId": "0198f101-1234-7000-8000-000000000001",
                    "requestTitle": "Ví dụ",
                    "senderId": "0198f101-1234-7000-8000-000000000001",
                    "senderName": "Ví dụ",
                    "senderRole": "ADMIN",
                    "message": "Ví dụ",
                    "messageType": "TEXT",
                    "isRead": false,
                    "createdAt": "2026-09-12T03:00:00Z",
                    "persisted": false
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "403": {
            "$ref": "#/components/responses/SecurityDenied"
          },
          "404": {
            "$ref": "#/components/responses/Error404"
          },
          "500": {
            "$ref": "#/components/responses/Error500"
          }
        },
        "parameters": [
          {
            "name": "requestId",
            "in": "path",
            "required": true,
            "description": "ID tài nguyên UUID.",
            "schema": {
              "type": "string",
              "format": "uuid",
              "example": "0198f101-1234-7000-8000-000000000001"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendChatMessageRequest"
              },
              "example": {
                "message": "Tôi đã kiểm tra lỗi đăng nhập.",
                "messageType": "TEXT"
              }
            }
          }
        }
      }
    },
    "/api/chat/messages/{messageId}/read": {
      "patch": {
        "tags": [
          "Chat"
        ],
        "summary": "Đánh dấu tin nhắn đã đọc",
        "operationId": "readMessage",
        "description": "Quyền: ADMIN hoặc client tạo request hoặc developer được giao. Không body. Mọi người có quyền xem request đều được đánh dấu; isRead là cờ chung.",
        "x-access": "ADMIN hoặc client tạo request hoặc developer được giao",
        "responses": {
          "200": {
            "description": "Thành công.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/readMessageResponse"
                },
                "example": {
                  "status": 200,
                  "message": "success",
                  "data": {
                    "id": "0198f101-1234-7000-8000-000000000001",
                    "requestId": "0198f101-1234-7000-8000-000000000001",
                    "requestTitle": "Ví dụ",
                    "senderId": "0198f101-1234-7000-8000-000000000001",
                    "senderName": "Ví dụ",
                    "senderRole": "ADMIN",
                    "message": "Ví dụ",
                    "messageType": "TEXT",
                    "isRead": false,
                    "createdAt": "2026-09-12T03:00:00Z",
                    "persisted": false
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "403": {
            "$ref": "#/components/responses/SecurityDenied"
          },
          "404": {
            "$ref": "#/components/responses/Error404"
          },
          "500": {
            "$ref": "#/components/responses/Error500"
          }
        },
        "parameters": [
          {
            "name": "messageId",
            "in": "path",
            "required": true,
            "description": "ID tài nguyên UUID.",
            "schema": {
              "type": "string",
              "format": "uuid",
              "example": "0198f101-1234-7000-8000-000000000001"
            }
          }
        ]
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Dán accessToken từ login; Authorization: Bearer <accessToken>."
      }
    },
    "schemas": {
...
      "ChatMessageResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "requestId": {
            "type": "string",
            "format": "uuid"
          },
          "requestTitle": {
            "type": "string"
          },
          "senderId": {
            "type": "string",
            "format": "uuid"
          },
          "senderName": {
            "type": "string",
            "nullable": true
          },
          "senderRole": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Role"
              }
            ],
            "nullable": true
          },
          "message": {
            "type": "string"
          },
          "messageType": {
            "type": "string"
          },
          "isRead": {
            "type": "boolean",
            "description": "Cờ đọc chung, không riêng theo người đọc."
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "persisted": {
            "type": "boolean",
            "description": "false: còn trong bộ nhớ; true: đã lưu DB."
          }
        }
      },
...
      "CompanyResponse": {
        "type": "object",
        "properties": {
          "companyId": {
            "type": "string",
            "nullable": true
          },
          "companyName": {
            "type": "string",
            "nullable": true
          },
          "address": {
            "type": "string",
            "nullable": true
          },
          "phone": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "MemberResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "email": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          },
          "phone": {
            "type": "string",
            "nullable": true
          },
          "companyId": {
            "type": "string",
            "nullable": true
          },
          "companyName": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string"
          },
          "isDeleted": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "RegisterRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "minLength": 1,
            "pattern": "\\S",
            "format": "email"
          },
          "password": {
            "type": "string",
            "minLength": 1,
            "pattern": "\\S",
            "format": "password",
            "writeOnly": true
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "pattern": "\\S"
          },
          "role": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Role"
              }
            ],
            "description": "Source hiện chấp nhận cả ADMIN, DEVELOPER, CLIENT trên endpoint đăng ký public."
          },
          "phone": {
            "type": "string"
          },
          "companyId": {
            "type": "string",
            "description": "Tùy chọn; trim khoảng trắng. Nếu khác rỗng phải tồn tại trong companies, nếu không trả 404."
          }
        },
        "required": [
          "email",
          "password",
          "name",
          "role"
        ]
      },
      "SendChatMessageRequest": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "minLength": 1,
            "pattern": "\\S",
            "maxLength": 10000
          },
          "messageType": {
            "type": "string",
            "maxLength": 20,
            "default": "TEXT",
            "description": "Chuỗi tùy ý, trim và uppercase; null/rỗng dùng TEXT. Không phải enum."
          }
        },
        "required": [
          "message"
        ]
      },
...
      "Role": {
        "type": "string",
        "enum": [
          "ADMIN",
          "DEVELOPER",
          "CLIENT"
        ]
      },
...
      "ApiError": {
        "type": "object",
        "required": [
          "status",
          "errorCode",
          "message"
        ],
        "properties": {
          "status": {
            "type": "integer"
          },
          "errorCode": {
            "type": "string",
            "enum": [
              "AUTH_401",
              "AUTH_403",
              "VAL_400",
              "BAD_400",
              "RES_404",
              "SYS_500"
            ]
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            },
            "description": "Lỗi theo field; data null bị bỏ khỏi envelope."
          }
        }
      },
...
      "registerMemberResponse": {
        "type": "object",
        "required": [
          "status",
          "message",
          "data"
        ],
        "properties": {
          "status": {
            "type": "integer",
            "enum": [
              200
            ],
            "description": "Status trong body, kể cả khi HTTP 201."
          },
          "message": {
            "type": "string",
            "example": "Member registered successfully"
          },
          "data": {
            "$ref": "#/components/schemas/MemberResponse"
          }
        }
      },
      "listMembersResponse": {
        "type": "object",
        "required": [
          "status",
          "message",
          "data"
        ],
        "properties": {
          "status": {
            "type": "integer",
            "enum": [
              200
            ],
            "description": "Status trong body, kể cả khi HTTP 201."
          },
          "message": {
            "type": "string",
            "example": "success"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MemberResponse"
            }
          }
        }
      },
      "getMemberResponse": {
        "type": "object",
        "required": [
          "status",
          "message",
          "data"
        ],
        "properties": {
          "status": {
            "type": "integer",
            "enum": [
              200
            ],
            "description": "Status trong body, kể cả khi HTTP 201."
          },
          "message": {
            "type": "string",
            "example": "success"
          },
          "data": {
            "$ref": "#/components/schemas/MemberResponse"
          }
        }
      },
      "listCompaniesResponse": {
        "type": "object",
        "required": [
          "status",
          "message",
          "data"
        ],
        "properties": {
          "status": {
            "type": "integer",
            "enum": [
              200
            ],
            "description": "Status trong body, kể cả khi HTTP 201."
          },
          "message": {
            "type": "string",
            "example": "success"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CompanyResponse"
            }
          }
        }
      },
      "getCompanyResponse": {
        "type": "object",
        "required": [
          "status",
          "message",
          "data"
        ],
        "properties": {
          "status": {
            "type": "integer",
            "enum": [
              200
            ],
            "description": "Status trong body, kể cả khi HTTP 201."
          },
          "message": {
            "type": "string",
            "example": "success"
          },
          "data": {
            "$ref": "#/components/schemas/CompanyResponse"
          }
        }
      },
...
      "chatInboxResponse": {
        "type": "object",
        "required": [
          "status",
          "message",
          "data"
        ],
        "properties": {
          "status": {
            "type": "integer",
            "enum": [
              200
            ],
            "description": "Status trong body, kể cả khi HTTP 201."
          },
          "message": {
            "type": "string",
            "example": "success"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ChatMessageResponse"
            }
          }
        }
      },
      "chatHistoryResponse": {
        "type": "object",
        "required": [
          "status",
          "message",
          "data"
        ],
        "properties": {
          "status": {
            "type": "integer",
            "enum": [
              200
            ],
            "description": "Status trong body, kể cả khi HTTP 201."
          },
          "message": {
            "type": "string",
            "example": "success"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ChatMessageResponse"
            }
          }
        }
      },
      "sendMessageResponse": {
        "type": "object",
        "required": [
          "status",
          "message",
          "data"
        ],
        "properties": {
          "status": {
            "type": "integer",
            "enum": [
              200
            ],
            "description": "Status trong body, kể cả khi HTTP 201."
          },
          "message": {
            "type": "string",
            "example": "success"
          },
          "data": {
            "$ref": "#/components/schemas/ChatMessageResponse"
          }
        }
      },
      "readMessageResponse": {
        "type": "object",
        "required": [
          "status",
          "message",
          "data"
        ],
        "properties": {
          "status": {
            "type": "integer",
            "enum": [
              200
            ],
            "description": "Status trong body, kể cả khi HTTP 201."
          },
          "message": {
            "type": "string",
            "example": "success"
          },
          "data": {
            "$ref": "#/components/schemas/ChatMessageResponse"
          }
        }
      }
    },
    "responses": {
      "Error400": {
        "description": "VAL_400: body/validation sai. BAD_400: nghiệp vụ sai hoặc không có quyền tài nguyên.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            },
            "examples": {
              "business": {
                "value": {
                  "status": 400,
                  "errorCode": "BAD_400",
                  "message": "Bad request"
                }
              },
              "validation": {
                "value": {
                  "status": 400,
                  "errorCode": "VAL_400",
                  "message": "Email cannot be blank",
                  "data": {
                    "email": "Email cannot be blank"
                  }
                }
              }
            }
          }
        }
      },
...
      "Error404": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            },
            "examples": {
              "business": {
                "value": {
                  "status": 404,
                  "errorCode": "RES_404",
                  "message": "Resource not found"
                }
              }
            }
          }
        }
      },
      "Error500": {
        "description": "System internal error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            },
            "examples": {
              "business": {
                "value": {
                  "status": 500,
                  "errorCode": "SYS_500",
                  "message": "System internal error"
                }
              }
            }
          }
        }
      },
      "SecurityDenied": {
        "description": "403: thiếu/token không hợp lệ hoặc không đủ role ở security filter; body do framework quyết định, có thể rỗng. AccessDeniedException đến controller advice trả ApiError với AUTH_403."
      }
    }
  }
}